% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_generation.R
\name{moves_generation}
\alias{moves_generation}
\title{Generowanie mozliwych rochow}
\usage{
moves_generation(stock_status, max_stock)
}
\arguments{
\item{stock_status}{Aktualny stan stada gracza}

\item{max_stock}{Wektor mowiacy ile zwierzat konkretnego typu znajduje sie w calej grze}
}
\value{
possible_moves Macierz wszystkich mozliwych ruchow
}
\description{
Funkcja \code{moves_generation()} wszytkie mozliwe ruchy gracza przy danym stadzie.
}
\author{
Agnieszka Ciepielewska, Mawek Wawreniuk, Pawel Gorniak
}

