% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{reproduction}
\alias{reproduction}
\title{Rozmnazanie zwierzat po rzucie}
\usage{
reproduction(die1_result, die2_result, stock_status, max_stock,
  zawsze_jeden_krolik = FALSE, co_zjada_wilk = c(1, 1, 1, 1, 0, 0, 0))
}
\arguments{
\item{die1_result}{String, bedacy wynikiem rzutu pierwsza kostka (tej z wilkiem)}

\item{die2_result}{String, bedacy wynikiem rzutu druga kostka (tej z lisem)}

\item{stock_status}{Wektor posiadanych przez gracza zwierzat}

\item{max_stock}{Wektor maksymalnej liczby zwierzat w stadzie}

\item{zawsze_jeden_krolik}{Wartosc binarna oznaczajaca czy gramy w wersje zasad w ktorej
zawsze mamy co najmniej jednego krolika}

\item{co_zjada_wilk}{wektor zer i jedynek oznaczajacy czy wilk zjada konkretny rodzaj zwierzecia czy nie
(duzy pies zawsze powinien dostawac 0, bo jest to rozpatrywane w przypadku gdy nie mamy duzego psa)}
}
\value{
stock_status Stan stada gracza po rozpatrzeniu rzutow kostkami
}
\description{
Funkcja przeprowadzajaca zmiany w stadzie gracza po rzucie kostkami,
wykorzystuje funkcje \code{update_stocku}
}
\author{
Agnieszka Ciepielewska, Marek Wawreniuk, Pawel Gorniak
}

