% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_rf.R
\name{strategia_rf}
\alias{strategia_rf}
\title{Strategia oparta na Random Forest}
\usage{
strategia_rf(stock_status, max_stock = c(60, 24, 20, 12, 6, 4, 2))
}
\arguments{
\item{stock_status}{Stan stada gracza przed ruchem}

\item{max_stock}{Wektor mowiacy ile zwierzat konkretnego typu znajduje sie w calej grze
domyslnie c(60, 24, 20, 12, 6, 4, 2)}
}
\value{
stock_status Stan stada gracza po ruchu
}
\description{
Funkcja \code{strategia_rf()} wybiera ruch gracza wedlug przewidywan modelu random forest wgranego do pakietu.
Model przewiduje, ktory ruch doprowadzi do najszybszego zwycieztwa. Wykorzystuje funkcje \code{moves_generation()},
zeby miec z czego wybierac ruch
}
\examples{
stan_stada <- c(2,4,1,0,0,0,0)
stan_stada <- strategia_rf(stan_stada)

}
\author{
Agnieszka Ciepielewska
}

