% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_EV.R
\name{strategy_EV}
\alias{strategy_EV}
\title{Strategia Expected Value}
\usage{
strategy_EV(stock_status, max_stock = c(krolik = 60, owca = 24, swinia = 20,
  krowa = 12, kon = 6, maly_pies = 4, duzy_pies = 2))
}
\description{
Strategia polega na wybraniu ruchu, dla ktorego wartosc
 oczekiwana liczona w krolikach jest najwieksza. Funkcja \code{moves_generation()} 
 generuje wszystkie mozliwe ruchy. Funkcja \code{get_the_best_state} zwraca ruch o 
 największej
}

