% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stocku.R
\name{update_stocku}
\alias{update_stocku}
\title{Aktualizacja stada}
\usage{
update_stocku(stock_status, max_stock, die1_result = "", die2_result = "")
}
\arguments{
\item{stock_status}{Wektor posiadanych przez gracza zwierzat}

\item{max_stock}{Wektor maksymalnej liczby zwierzat w stadzie}

\item{die1_result}{String, bedacy wynikiem rzutu pierwsza kostka (tej z wilkiem), domyslnie pusty string}

\item{die2_result}{String, bedacy wynikiem rzutu druga kostka (tej z lisem), domyslnie pusty string}
}
\value{
stock_status Zaktualizowany stan stada gracza
}
\description{
Funkcja pomocnicza wykorzystywana przez funkcje \code{reproduction()} do aktualizowania stanu stada gracza
}
\author{
Agnieszka Ciepielewska
}

