% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win.R
\name{win}
\alias{win}
\title{Czy gracz wygral}
\usage{
win(stock_status, warunek_zwycieztwa = c(1, 1, 1, 1, 1, 0, 0))
}
\arguments{
\item{stock_status}{Wektor aktualnego stanu stada gracza}

\item{warunek_zwycieztwa}{Jakie zwierzeta musi miec gracz aby wygrac}
}
\value{
Funkcja zwraca \code{TRUE} jesli gracz spelnia warunki zwyciestwa i \code{FALSE} w przeciwnym przypadku
}
\description{
Funkcja sprawdzajaca czy w danym momencie gracz ma odpowiednie zwierzeta, aby wygrac
}
\author{
Agnieszka Ciepielewska, Marek Wawreniuk, Pawel Gorniak
}

